requirements.txt
streamlit==0.60.0
numpy==1.17.2
transformers==2.10.0
newspaper3k==0.2.8
nltk==3.5
vncorenlp==1.0.3